+++ include/Solver.h
@@ -20,8 +20,8 @@ class Solver
 {
     private:
         Solver() {};
-        float** lpToTableau(LinearProgram* lp, int* numDecisionVars, int* numConstraints);
-        void Pivot(float** tableau, int* pivotRow, int* pivotCol, int* numDecisionVars, int* numConstraints);
+        double** lpToTableau(LinearProgram* lp, int* numDecisionVars, int* numConstraints);
+        void Pivot(double** tableau, int* pivotRow, int* pivotCol, int* numDecisionVars, int* numConstraints);
         // Don't implement these, they prevent copies from being made.
         Solver(Solver const& copy);
         Solver& operator=(Solver const& copy);
