+++ src/main.cpp
@@ -12,7 +12,7 @@ int main(void)
     pTestProblem->AddConstraint("3 4 2 8");
     LPSolution answer = Solver::getInstance().SimplexSolve(pTestProblem);
     std::cout << "answer error code = " << answer.GetErrorCode() << std::endl;
-    std::vector<float> answervals = answer.GetOptimalValues();
+    std::vector<double> answervals = answer.GetOptimalValues();
     std::cout << "answer values: " << answervals[0] << std::endl; 
     std::cout << "answer values: " << answervals[1] << std::endl; 
     std::cout << "answer values: " << answervals[2] << std::endl;
